Remove db exception caused by attendee deletion
When deleting an attendee, Django emulates the database deletion logic for rows referencing each other. Django will detect that seletions must be deleted before deleting an attendee object. However, signals were used to detect selection deletion and create a log entry referencing the soon-to-be-deleted attendee - causing an error in MySQL.